Always align characters to bytes
Show ascii escape characters rather than rendering control characters directly,
and similarly show unicode codepoints for anything outside of ascii.

The second decision is not ideal, because if there are printable non ascii
characters you should be able to see what they are. But this allows me to
postpone dealing with a lot of complexity around how fonts render unicode, for example:
  - whitespace and halfwidth vs fullwidth characters will complicate my attempts to align bytes and characters
  - combining characters or RTL markers will mess up output that comes after it
  - where a grapheme cluster is made up of multiple printable characters, showing individual characters might not be the most useful thing to display