Add line wrapping
If the representation of the input would exceed the width of the terminal, break it up into lines
that will fit.

This is to ensure that the bytes/chars representations continue to line up properly.

This commit also changes the formatting of the 'output' string - rather than prefixing this, put it onto its own line
and highlight the whole thing to visually distinguish it from the character/byte information. Otherwise input that
contains newlines will look a bit weird, since the prefix causes the first line to be indented.