Change parse_input to return a vector of bytes
Vec<u8> means we can interpret the text using a range of encodings, not just utf8,
which is the only one supported at the moment.

The method into_vec() on OsString only works on Unix, though: https://doc.rust-lang.org/std/ffi/struct.OsString.html

On Unix you can think of strings as a series of bytes which are conventionally
utf-8. If the user passes in strings as arguments we will display the same
series of bytes.

On windows, rust is actually taking a stream of 16 bit code units, and then reencoding
using WTF-8 (https://simonsapin.github.io/wtf-8/). Which means that the program doesn't actually
see the operating system's representation of the string directly.

Although this reencoding is lossless, I don't know how to make this work
reliabily, and I'm not testing on windows at the moment, so I'm not going to
bother making this cross platform for now.

This shouldn't affect the case where zero arguments are passed in and we read from stdin - in
that case we can just read the file into a Vec<u8> directly and do not need OsStrings.