Render the correct bytes when displaying ISO 8859-1
Make the rendering code encoding-independent rather than assuming utf8.

This is still really dodgy because it re-encodes the decoded string
character by character, which is potentially lossy. For example, if
the original decoding required replacement, it will re-encode that
replacement character, so we will display that instead of the original
bytes that we couldn't decode.

The reason I'm doing this is to be able to associate parts of the byte string
with the individual characters we decoded.

I want to find a better way of doing this but I think this code needs to be
refactored first. There are also a lot of edge cases that are currently
unhandled.