Introduce DecodedUnit type
A DecodedUnit is either a character or a byte we failed to decode.

If there is a sequence of invalid code units in a row then they there
is a sequence of DecodedUnit::Error values.

Introducing this type allows us to preserve the original bytes to
display to the user, even if they couldn't be decoded.