Rename DecodedUnit to Atom
Similarly, rename DecodedString.characters -> DecodedString.atoms. This
ensures we are referring to this thing in a consistent way.

"Character" is a very overloaded term, and it no longer makes sense for
DecodedString to have a field called characters, given that members of vector
can represent invalid code units that were not decoded to a character.

I couldn't think of a good name for this thing, but I've gone with Atom because
it represents the smallest division of a string we care about, whether that
corresponds to a valid character or not.

DecodedCharacter is also a crappy name, but I've left that alone for now, as I
couldn't think of anything better.