Handle *some* truncated inputs
Strings can be truncated, leaving incomplete sequences of bytes.
In this case the error can only be detected by calling the `raw_finish`.
method of the `RawDecoder` trait.

We can handle the case where only the last byte of the input was incomplete,
but the error is not telling us exactly how many bytes were invalid.
So if I have a string full of emoji then I don't know how many bytes
were already decoded.